+++ client/testing.go
@@ -9,7 +9,7 @@ import (
 	"github.com/hashicorp/nomad/client/fingerprint"
 	"github.com/hashicorp/nomad/command/agent/consul"
 	"github.com/hashicorp/nomad/helper/testlog"
-	"github.com/hashicorp/nomad/plugins/shared/catalog"
+	"github.com/hashicorp/nomad/pluginutils/catalog"
 	"github.com/hashicorp/nomad/plugins/shared/singleton"
 	testing "github.com/mitchellh/go-testing-interface"
 )
