+++ pluginutils/catalog/register.go
