+++ pluginutils/catalog/register_linux.go
